+++ pkg/sass-parser/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.4.33
+
+* No user-visible changes.
+
 ## 0.4.32
 
 * No user-visible changes.